convert generic drf example on using list and detail views
drf-gis sublclasses the drf modelview, not the hyperlinkedmodelview, so we'll have to to a bit more configuration.  So far, we are now returning the list view which has the geojson for all alder districts, and the detail view, which has the geojson for a single district.  Hyperlinking has not yet been added.